修改告警
issue:https://gitee.com/openharmony/arkcompiler_ets_runtime/issues/I9KXYD

+++ ecmascript/compiler/bytecode_circuit_builder.h
@@ -550,7 +550,7 @@ public:
         int32_t preConstpoolId = 0;
         while (gateAcc_.GetOpCode(frameArgs) != OpCode::CIRCUIT_ROOT) {
             preConstpool = gateAcc_.GetValueIn(frameArgs, static_cast<size_t>(FrameArgIdx::CONST_POOL));
-            preConstpoolId = gateAcc_.GetConstpoolId(preConstpool);
+            preConstpoolId = static_cast<int32_t>(gateAcc_.GetConstpoolId(preConstpool));
             if (preConstpoolId == constpoolId) {
                 return preConstpool;
             }
